# improve elementwise performance.